# Optimise WAL loading by caching min-time