Fixes #8638

Profiling shows that computing the last timestamp of mapped chunks inside `append()` takes ~10% of CPU time, so compute and store it once per series.

Edit: also fix some issues in the benchmark.