@bboreham 2021-03-25
As far as I can see `BenchmarkLoadWAL()` sets up a WAL with no mmapped chunks, so the code I noted in the profile is not executed.

I think I need some advice on how to set up a benchmark with both WAL samples and mmapped chunks.