@bboreham 2021-07-19
I spent some time drilling into the behaviour of the code, figured out how to create some mmapped chunks in `BenchmarkLoadWAL()`.
When I profile the benchmark, much of the time is spent in the Go scheduler; is that expected?
I think I found one flaw in the benchmark; I'm interested to hear whether it has been closely examined before.