@codesome 2021-07-19
> I'm interested to hear whether it has been closely examined before.

Prolly not