@bboreham 2021-07-20
I have added a new option to `BenchmarkLoadWAL()` to simulate mmapped chunks, and a few more commits to address some items noted at #9101.

Headline benchmark numbers say not much has changed:
```
name                                                                                                    old time/op    new time/op    delta
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-4         582ms ±20%     537ms ± 7%    ~     (p=0.222 n=5+5)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-4         866ms ±17%     814ms ±10%    ~     (p=0.310 n=5+5)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-4         329ms ± 2%     326ms ± 5%    ~     (p=0.690 n=5+5)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-4     3.26s ± 5%     3.21s ± 3%    ~     (p=0.421 n=5+5)

name                                                                                                    old alloc/op   new alloc/op   delta
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-4        52.1MB ± 0%    52.1MB ± 0%  +0.07%  (p=0.008 n=5+5)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-4         260MB ± 8%     254MB ± 7%    ~     (p=0.310 n=5+5)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-4        51.5MB ± 2%    51.9MB ± 3%    ~     (p=0.841 n=5+5)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-4     351MB ± 0%     361MB ± 1%  +2.69%  (p=0.008 n=5+5)

name                                                                                                    old allocs/op  new allocs/op  delta
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-4          586k ± 0%      586k ± 0%    ~     (p=0.421 n=5+5)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-4         2.49M ± 0%     2.49M ± 0%    ~     (p=0.310 n=5+5)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-4          487k ± 0%      487k ± 0%    ~     (p=0.905 n=5+4)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-4     3.73M ± 0%     3.73M ± 0%    ~     (p=0.841 n=5+5)
```

however if we look at the amount of CPU used during the test with simulated mmapped chunks we see about a 10% improvement, in line with what the profile showed at #8638:
```
before: 13.88user 2.82system 0:07.31elapsed 228%CPU (0avgtext+0avgdata 275464maxresident)k
after:  12.39user 2.26system 0:07.02elapsed 208%CPU (0avgtext+0avgdata 275956maxresident)k
```