@bboreham 2021-07-26
I started to re-word the comment, then realised the whole basis for this cache is incorrect.

"Mitigate lock contention in getByID", it said, but the locks are sharded 32,768 ways so with any likely number of workers there is no contention.

Removing `refSeries` altogether makes my test go about as fast as adding `minTime` did.  The real problem must be somewhere else.