@cstyan 2021-07-27
> Removing refSeries altogether makes my test go about as fast as adding minTime did. The real problem must be somewhere else.

Interesting.